Created a 'menus' file to simplify things.
