+++ lib/helpers.js
@@ -9,6 +9,15 @@ const isOnline = require('is-online');
 
 const conf = new Configstore('github-star-transfer');
 
+const {
+	editUserCredentialsQuestions1,
+	exitMenuQuestions,
+	settingsMenuQuestions,
+	setupMenuQuestions1,
+	setupStartQuestions,
+	viewUserAccountQuestions
+} = require('./menus');
+
 /**
  * Main Menu Questions Object.
  */
@@ -19,106 +28,6 @@ var mainMenuQuestions = {
 	choices: ['Copy Stars', separator(), 'Settings', 'Exit']
 };
 
-/**
- * Settings Menu Questions Object
- */
-const settingsMenuQuestions = {
-	type: 'list',
-	name: 'settings',
-	message: 'SETTINGS',
-	choices: [
-		'Edit User Credentials',
-		'View User Accounts',
-		separator(),
-		'Main Menu',
-		'Exit'
-	]
-};
-
-/**
- * Exit Menu Questions Object
- */
-const exitMenuQuestions = {
-	type: 'list',
-	name: 'exit',
-	message: 'Are you sure you want to quit GitHub Star Transfer?',
-	choices: ['Yes', separator(), 'No (Back)']
-};
-
-/**
- * 1st Set of Setup Menu Questions
- */
-const setupMenuQuestions1 = [
-	{
-		name: 'user1Username',
-		type: 'input',
-		message: 'Enter the username for the account to transfer from:',
-		validate: function(value) {
-			if (value.length) {
-				return true;
-			} else {
-				return "Please enter the account's username.";
-			}
-		}
-	},
-	{
-		name: 'user1Password',
-		type: 'password',
-		message: 'Enter the password for the account to transfer from:',
-		validate: function(value) {
-			if (value.length) {
-				return true;
-			} else {
-				return "Please enter the account's password.";
-			}
-		}
-	},
-	{
-		name: 'user2Username',
-		type: 'input',
-		message: 'Enter the username for the account to transfer to:',
-		validate: function(value) {
-			if (value.length) {
-				return true;
-			} else {
-				return "Please enter the account's username.";
-			}
-		}
-	},
-	{
-		name: 'user2Password',
-		type: 'password',
-		message: 'Enter the password for the account to transfer to:',
-		validate: function(value) {
-			if (value.length) {
-				return true;
-			} else {
-				return "Please enter the account's password.";
-			}
-		}
-	}
-];
-
-/**
- * Edit User Credentials Question Object
- */
-const editUserCredentialsQuestions1 = {
-	type: 'list',
-	name: 'credentials1',
-	message: 'Which account credentials would you like to edit?',
-	choices: [
-		'Account 1 - (Transfer FROM)',
-		'Account 2 - (Transfer TO)',
-		separator(),
-		'Back'
-	]
-};
-
-const viewUserAccountQuestions = {
-	type: 'confirm',
-	name: 'viewAccounts',
-	message: 'Return to main menu?'
-};
 
 /**
  * Displays a title screen in a pastel color.
@@ -454,5 +363,7 @@ module.exports = {
 	printTitle,
 	timeout,
 	createMenu,
-	separator
+	separator,
+	setupStartQuestions,
+	mainMenuQuestions
 };
